Refactor conditions
